backend/requirements.txt
Flask==2.3.2
rembg==2.0.30      # if you use rembg
Pillow==10.0.0
